Add FunctionModel for handling functions
This model type uses a Pydantic model to validate function
arguments and return values, exposing functions to pass in and
retrieve arguments, and to call functions using the stored
arguments.

This is a feature that has been asked for previously in
pydantic/pydantic#1391 but which was possible through v1 methods
which are not present in v2 and so that issue is already closed.